links added

+++ src/Utils/ApiData.php
@@ -21,7 +21,7 @@ class ApiData
         }
     }
 
-    public function getPrice(array $mealNames,array $sizes,int $i):int
+    public function getPrice(array $mealNames,array $sizes,int $i):float
     {
         $mealName = $mealNames[$i];
         $size = $sizes[$i];
@@ -33,14 +33,14 @@ class ApiData
             return $price;
     }
 
-    private function countPricexQty(array $mealNames,array $sizes,array $qtys,int $i):int{
+    private function countPricexQty(array $mealNames,array $sizes,array $qtys,int $i):float{
         $qty = $qtys[$i];
         $price = $this->getPrice($mealNames,$sizes,$i);
         return $price * $qty;
 
     }
 
-    public function countTotalPrice(array $mealNames,array $sizes,array $qtys):int{
+    public function countTotalPrice(array $mealNames,array $sizes,array $qtys):float{
         $valueArr = [];
 
         foreach ($mealNames as $key=>$value) {
